support localhost and ip addresses
