+++ packages/driver/cypress/integration/e2e/multi-domain/multi_domain_spec.ts
@@ -145,100 +145,6 @@ describe('multi-domain', { experimentalSessionSupport: true, experimentalMultiDo
   })
 
   describe('errors', () => {
-    // @ts-ignore
-    it('errors if experimental flag is not enabled', { experimentalMultiDomain: false }, (done) => {
-      cy.on('fail', (err) => {
-        expect(err.message).to.equal('`cy.switchToDomain()` requires enabling the experimentalMultiDomain flag')
-
-        done()
-      })
-
-      // @ts-ignore
-      cy.switchToDomain()
-    })
-
-    it('errors if passed a non-string for the domain argument', (done) => {
-      cy.on('fail', (err) => {
-        expect(err.message).to.equal('`cy.switchToDomain()` requires the first argument to be a string representing a domain that does not not include any subdomains. You passed: ``')
-
-        done()
-      })
-
-      // @ts-ignore
-      cy.switchToDomain()
-    })
-
-    it('fails on a bad domain name', (done) => {
-      cy.on('fail', (err) => {
-        expect(err.message).to.equal('`cy.switchToDomain()` requires the first argument to be a string representing a domain that does not not include any subdomains. You passed: `foobar`')
-
-        done()
-      })
-
-      cy.switchToDomain('foobar', () => undefined)
-    })
-
-    it('fails on a domain name with a sub domain', (done) => {
-      cy.on('fail', (err) => {
-        expect(err.message).to.equal('`cy.switchToDomain()` requires the first argument to be a string representing a domain that does not not include any subdomains. You passed: `eu.foobar.com`')
-
-        done()
-      })
-
-      cy.switchToDomain('eu.foobar.com', () => undefined)
-    })
-
-    it('errors passing non-array to callback function', (done) => {
-      cy.on('fail', (err) => {
-        expect(err.message).to.equal('`cy.switchToDomain()` requires the \'data\' argument to be an array. You passed: `foo`')
-
-        done()
-      })
-
-      // @ts-ignore
-      cy.switchToDomain('foobar.com', 'foo', () => {})
-    })
-
-    it('errors if passed a non-serializable data value', (done) => {
-      cy.on('fail', (err) => {
-        expect(err.message).to.include('data argument specified is not serializable')
-
-        if (Cypress.browser.family === 'chromium') {
-          expect(err.message).to.include('HTMLDivElement object could not be cloned')
-        } else if (Cypress.browser.family === 'firefox') {
-          expect(err.message).to.include('The object could not be cloned')
-        }
-
-        done()
-      })
-
-      const el = document.createElement('div')
-
-      cy.switchToDomain('foobar.com', ['foo', '1', el], () => {})
-    })
-
-    it('errors if last argument is absent', (done) => {
-      cy.on('fail', (err) => {
-        expect(err.message).to.equal('`cy.switchToDomain()` requires the last argument to be a function. You passed: ``')
-
-        done()
-      })
-
-      // @ts-ignore
-      cy.switchToDomain('foobar.com')
-    })
-
-    it('errors if last argument is not a function', (done) => {
-      cy.on('fail', (err) => {
-        expect(err.message).to.equal('`cy.switchToDomain()` requires the last argument to be a function. You passed: `{}`')
-
-        done()
-      })
-
-      // @ts-ignore
-      cy.switchToDomain('foobar.com', {})
-    })
-
     // TODO: Proper stack trace printing still needs to be addressed here
     it('propagates secondary domain errors to the primary that occur within the test', () => {
       return new Promise((resolve) => {
