+++ packages/driver/package.json
@@ -54,6 +54,7 @@
     "errorhandler": "1.5.1",
     "eventemitter2": "6.4.2",
     "express": "4.17.1",
+    "is-ip": "4.0.0",
     "is-valid-domain": "0.0.20",
     "is-valid-hostname": "1.0.1",
     "jquery": "3.1.1",
