+++ packages/driver/src/cy/multi-domain/validator.ts
@@ -2,6 +2,8 @@ import type $Log from '../../cypress/log'
 import $utils from '../../cypress/utils'
 import $errUtils from '../../cypress/error_utils'
 import isValidDomain from 'is-valid-domain'
+import { isIP } from 'is-ip'
+import { isString } from 'lodash'
 
 export class Validator {
   log: typeof $Log
@@ -13,8 +15,7 @@ export class Validator {
   }
 
   validate ({ callbackFn, data, domain }) {
-    if (!isValidDomain(domain, { allowUnicode: true, subdomain: false })) {
-
+    if (!isString(domain) || domain !== 'localhost' && !isIP(domain) && !isValidDomain(domain, { allowUnicode: true, subdomain: false })) {
       this.onFailure()
 
       $errUtils.throwErrByPath('switchToDomain.invalid_domain_argument', {
